src/PDFTOJSON2FunctionApp/requirements.txt
azure-functions
azure-ai-formrecognizer
azure-ai-ml
azure-storage-blob
openai
python-dotenv
requests
